requirements.txt
pydantic
pydantic_core
openai
google-genai
aiohttp
python-dotenv
